Rewrite to minimize changes to DefaultFareServiceImpl. No configuration necessary.
