+++ src/main/java/org/opentripplanner/routing/impl/DefaultFareServiceImpl.java
@@ -146,14 +146,6 @@ class FareAndId {
     }
 }
 
-/** Allow a FareService to override the default mode of adding fares together for subsets of rides */
-interface FareAdditiveStrategy {
-    /**
-     * Given sets of rides and associated costs, add the costs together.
-     */
-    float addFares(List<Ride> ride0, List<Ride> ride1, float cost0, float cost1);
-}
-
 /**
  * This fare service module handles the cases that GTFS handles within a single feed.
  * It cannot necessarily handle multi-feed graphs, because a rule-less fare attribute
@@ -169,8 +161,6 @@ public class DefaultFareServiceImpl implements FareService, Serializable {
 
     private static final Logger LOG = LoggerFactory.getLogger(DefaultFareServiceImpl.class);
 
-    private FareAdditiveStrategy fareAdditiveStrategy;
-
     /** For each fare type (regular, student, etc...) the collection of rules that apply. */
     protected Map<FareType, Collection<FareRuleSet>> fareRulesPerType;
 
@@ -180,10 +170,6 @@ public class DefaultFareServiceImpl implements FareService, Serializable {
 
     public void addFareRules(FareType fareType, Collection<FareRuleSet> fareRules) {
         fareRulesPerType.put(fareType, new ArrayList<>(fareRules));
-    }
-
-    public void setFareAdditiveStrategy(FareAdditiveStrategy fareAdditiveStrategy) {
-        this.fareAdditiveStrategy = fareAdditiveStrategy;
     }
 
     protected List<Ride> createRides(GraphPath path) {
@@ -276,14 +262,8 @@ public class DefaultFareServiceImpl implements FareService, Serializable {
                 r.resultTable[j][j + i] = cost;
                 r.fareIds[j][j + i] = best.fareId;
                 for (int k = 0; k < i; k++) {
-                    float via;
-                    if (fareAdditiveStrategy == null) {
-                        via = r.resultTable[j][j + k] + r.resultTable[j + k + 1][j + i];
-                    }
-                    else {
-                        via = fareAdditiveStrategy.addFares(rides.subList(j, j + k + 1), rides.subList(j + k + 1, j + i + 1),
-                                r.resultTable[j][j + k], r.resultTable[j + k + 1][j + i]);
-                    }
+                    float via = addFares(rides.subList(j, j + k + 1), rides.subList(j + k + 1, j + i + 1),
+                            r.resultTable[j][j + k], r.resultTable[j + k + 1][j + i]);
                     if (r.resultTable[j][j + i] > via) {
                         r.resultTable[j][j + i] = via;
                         r.endOfComponent[j] = j + i;
@@ -293,6 +273,10 @@ public class DefaultFareServiceImpl implements FareService, Serializable {
             }
         }
         return r;
+    }
+
+    protected float addFares(List<Ride> ride0, List<Ride> ride1, float cost0, float cost1) {
+        return cost0 + cost1;
     }
 
     protected float getLowestCost(FareType fareType, List<Ride> rides,
